chore: delete list/business-example
